# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Investments Belfort API
  version: "1.0"
  description: API contract for the crypto investment analysis platform.
servers:
  - url: http://localhost:8000
    description: Local gateway
paths:
  /ohlcv/{symbol}:
    get:
      operationId: getOhlcv
      summary: Get OHLCV candlestick data
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: tf
          in: query
          schema:
            $ref: "#/components/schemas/Timeframe"
        - name: limit
          in: query
          schema:
            type: integer
            default: 500
      responses:
        "200":
          description: OHLCV data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OhlcvResponse"
  /analysis/{symbol}:
    get:
      operationId: getAnalysis
      summary: Get technical analysis for a symbol
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: tf
          in: query
          schema:
            $ref: "#/components/schemas/Timeframe"
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Technical analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResponse"
  /analysis/{symbol}/chart-markers:
    get:
      operationId: getChartMarkers
      summary: Get chart annotation markers for viewport
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: tf
          in: query
          schema:
            $ref: "#/components/schemas/Timeframe"
        - name: history
          in: query
          schema:
            $ref: "#/components/schemas/HistoryDepth"
        - name: from
          in: query
          schema:
            type: integer
        - name: to
          in: query
          schema:
            type: integer
        - name: min_confidence
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 1
            default: 0.35
        - name: pattern_status
          in: query
          schema:
            $ref: "#/components/schemas/ChartPatternStatusFilter"
        - name: pattern_direction
          in: query
          schema:
            $ref: "#/components/schemas/ChartPatternDirectionFilter"
        - name: categories
          in: query
          description: Comma-separated pattern categories (e.g. candles,trend) or omit for all
          schema:
            type: string
        - name: max_markers
          in: query
          schema:
            type: integer
            minimum: 5
            maximum: 200
            default: 80
      responses:
        "200":
          description: Chart markers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartMarker"
  /analysis/{symbol}/chart-overlays:
    get:
      operationId: getChartOverlays
      summary: Projection path, EMA, levels and trade setup for chart
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: tf
          in: query
          schema:
            $ref: "#/components/schemas/Timeframe"
        - name: history
          in: query
          schema:
            $ref: "#/components/schemas/HistoryDepth"
        - name: projection_bars
          in: query
          schema:
            type: integer
            minimum: 4
            maximum: 48
            default: 16
        - name: min_level_strength
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 1
            default: 0
        - name: max_levels_per_side
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
        - name: min_confluence
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
            default: 0
      responses:
        "200":
          description: Chart overlay bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartOverlaysResponse"
...
  /backtest/{symbol}:
    get:
      operationId: getBacktest
      summary: Get backtest results for a symbol and strategy
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: strategy
          in: query
          schema:
            type: string
            default: "v1"
        - name: tf
          in: query
          schema:
            $ref: "#/components/schemas/Timeframe"
      responses:
        "200":
          description: Backtest results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestResponse"
  /backtest/{symbol}/refresh:
    post:
      operationId: refreshBacktest
      summary: Trigger a backtest refresh (quick or full grid)
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: tf
          in: query
          schema:
            $ref: "#/components/schemas/Timeframe"
        - name: mode
          in: query
          schema:
            type: string
            enum: ["quick", "full"]
            default: "quick"
      responses:
        "200":
          description: Job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
  /analysis/{symbol}/patterns/all:
    get:
      operationId: getPatternsAll
      summary: Get all patterns with active/recent/inactive status
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: tf
          in: query
          schema:
            $ref: "#/components/schemas/Timeframe"
        - name: history
          in: query
          schema:
            $ref: "#/components/schemas/HistoryDepth"
      responses:
        "200":
          description: All patterns with status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatternsAllResponse"
  /analysis/{symbol}/refresh:
    post:
      operationId: refreshAnalysis
      summary: Trigger analysis data refresh
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: tf
          in: query
          schema:
            $ref: "#/components/schemas/Timeframe"
      responses:
        "200":
          description: Job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
...
  /jobs/{job_id}:
    get:
      operationId: getJob
      summary: Get job status and progress
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
  /jobs:
    get:
      operationId: listJobs
      summary: List recent jobs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of recent jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobResponse"
...
components:
  schemas:
    Timeframe:
      type: string
      enum: ["1h", "4h", "1D", "1W"]
    HistoryDepth:
      type: string
      enum: ["2y", "5y", "max"]
      default: "5y"
    TrendDirection:
      type: string
      enum: ["bullish", "bearish", "neutral"]
    PatternDirection:
      type: string
      enum: ["bull", "bear", "neutral"]
    PatternCategory:
      type: string
      enum: ["candles", "chart_patterns", "trend", "structure", "indicators", "levels"]
...
    OhlcvCandle:
      type: object
      required: [time, open, high, low, close, volume]
      properties:
        time:
          type: integer
          description: Unix timestamp in seconds
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
    OhlcvResponse:
      type: object
      required: [symbol, tf, candles]
      properties:
        symbol:
          type: string
        tf:
          $ref: "#/components/schemas/Timeframe"
        candles:
          type: array
          items:
            $ref: "#/components/schemas/OhlcvCandle"
    PatternItem:
      type: object
      required: [id, name, category, direction, bar_time, confidence, active]
      properties:
        id:
          type: string
        name:
          type: string
        category:
          $ref: "#/components/schemas/PatternCategory"
        direction:
          $ref: "#/components/schemas/PatternDirection"
        bar_time:
          type: integer
        confidence:
          type: number
          minimum: 0
          maximum: 1
        active:
          type: boolean
        description:
          type: string
        tf:
          $ref: "#/components/schemas/Timeframe"
        highlight:
          type: boolean
          default: false
    IndicatorItem:
      type: object
      required: [name, value, signal]
      properties:
        name:
          type: string
        value:
          type: number
        signal:
          $ref: "#/components/schemas/TrendDirection"
        zone:
          type: string
          enum: ["overbought", "oversold", "neutral"]
        description:
          type: string
    Level:
      type: object
      required: [price, type, strength]
      properties:
        price:
          type: number
        type:
          type: string
          enum: ["support", "resistance"]
        strength:
          type: number
          minimum: 0
          maximum: 1
    TradeSetup:
      type: object
      properties:
        entry:
          type: number
        stop_loss:
          type: number
        take_profit:
          type: number
        risk_reward:
          type: number
        direction:
          $ref: "#/components/schemas/TrendDirection"
    PatternGroup:
      type: object
      required: [category, count, items]
      properties:
        category:
          $ref: "#/components/schemas/PatternCategory"
        count:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PatternItem"
    AnalysisSummary:
      type: object
      required: [trend, confluence_score, highlights]
      properties:
        trend:
          $ref: "#/components/schemas/TrendDirection"
        trend_strength:
          type: integer
          minimum: 0
          maximum: 100
        confluence_score:
          type: integer
          minimum: 0
          maximum: 100
        highlights:
          type: array
          maxItems: 5
          items:
            $ref: "#/components/schemas/PatternItem"
        trade_setup:
          $ref: "#/components/schemas/TradeSetup"
        timeframe_analysis:
          type: object
          additionalProperties:
            type: object
            properties:
              trend:
                $ref: "#/components/schemas/TrendDirection"
              strength:
                type: integer
    AnalysisResponse:
      type: object
      required: [symbol, tf, updated_at, summary, groups, indicators, levels]
      properties:
        symbol:
          type: string
        tf:
          $ref: "#/components/schemas/Timeframe"
        updated_at:
          type: string
          format: date-time
        summary:
          $ref: "#/components/schemas/AnalysisSummary"
        groups:
          type: array
          items:
            $ref: "#/components/schemas/PatternGroup"
        indicators:
          type: array
          items:
            $ref: "#/components/schemas/IndicatorItem"
        levels:
          type: array
          items:
            $ref: "#/components/schemas/Level"
        pagination:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
    ChartPatternStatusFilter:
      type: string
      enum: [highlight, active, recent, active_recent, all]
      default: active_recent
    ChartPatternDirectionFilter:
      type: string
      enum: [all, bull, bear]
      default: all
    ChartMarker:
      type: object
      required: [time, position, shape, color, text]
      properties:
        time:
          type: integer
        position:
          type: string
          enum: ["aboveBar", "belowBar", "inBar"]
        shape:
          type: string
          enum: ["arrowUp", "arrowDown", "circle", "square"]
        color:
          type: string
        text:
          type: string
        pattern_id:
          type: string
    ChartProjectionPoint:
      type: object
      required: [time, value]
      properties:
        time:
          type: integer
        value:
          type: number
    ChartProjection:
      type: object
      required: [direction, target_price, bars, path, candles]
      properties:
        direction:
          $ref: "#/components/schemas/TrendDirection"
        target_price:
          type: number
        bars:
          type: integer
        path:
          type: array
          items:
            $ref: "#/components/schemas/ChartProjectionPoint"
        candles:
          type: array
          items:
            $ref: "#/components/schemas/OhlcvCandle"
    EmaSeries:
      type: object
      required: [period, points]
      properties:
        period:
          type: integer
        points:
          type: array
          items:
            $ref: "#/components/schemas/ChartProjectionPoint"
    ChartOverlaysResponse:
      type: object
      required: [symbol, tf, trend, confluence_score, ema, levels]
      properties:
        symbol:
          type: string
        tf:
          $ref: "#/components/schemas/Timeframe"
        trend:
          $ref: "#/components/schemas/TrendDirection"
        confluence_score:
          type: integer
        projection:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ChartProjection"
        ema:
          type: array
          items:
            $ref: "#/components/schemas/EmaSeries"
        levels:
          type: array
          items:
            $ref: "#/components/schemas/Level"
        trade_setup:
          $ref: "#/components/schemas/TradeSetup"
...
    EquityPoint:
      type: object
      required: [date, value]
      properties:
        date:
          type: string
        value:
          type: number
    BacktestResponse:
      type: object
      required: [symbol, strategy, tf, metrics, updated_at, status]
      properties:
        symbol:
          type: string
        strategy:
          type: string
        tf:
          $ref: "#/components/schemas/Timeframe"
        status:
          type: string
          enum: [empty, running, ready]
          description: empty = no results yet, running = job in progress, ready = results available
        metrics:
          type: object
          required: [total_trades, win_rate, profit_factor, max_drawdown, sharpe]
          properties:
            total_trades:
              type: integer
            win_rate:
              type: number
            profit_factor:
              type: number
            max_drawdown:
              type: number
            sharpe:
              type: number
            avg_return:
              type: number
        equity_curve:
          type: array
          items:
            $ref: "#/components/schemas/EquityPoint"
        updated_at:
          type: string
          format: date-time
          nullable: true
        note:
          type: string
        run_summary:
          type: object
          properties:
            last_run_at:
              type: string
              format: date-time
              nullable: true
            patterns_with_results:
              type: integer
            total_combos_saved:
              type: integer
            grid_mode:
              type: string
              enum: [quick, full]
              nullable: true
            selection_metric:
              type: string
        selection:
          type: object
          nullable: true
          properties:
            pattern_id:
              type: string
            pattern_name:
              type: string
            pattern_category:
              type: string
              nullable: true
            sort_metric:
              type: string
            run_at:
              type: string
              format: date-time
              nullable: true
            params:
              type: object
              additionalProperties: true
        active_job:
          $ref: "#/components/schemas/BacktestJobSummary"
        last_job:
          $ref: "#/components/schemas/BacktestJobSummary"
    BacktestJobSummary:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        mode:
          type: string
          enum: [quick, full]
        status:
          type: string
          enum: [pending, running, done, failed]
        created_at:
          type: string
          format: date-time
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        progress_current:
          type: integer
        progress_total:
          type: integer
        progress_message:
          type: string
        result:
          nullable: true
...
    PatternAllItem:
      type: object
      required: [id, name, category, status, direction]
      properties:
        id:
          type: string
        name:
          type: string
        category:
          $ref: "#/components/schemas/PatternCategory"
        status:
          type: string
          enum: ["active", "recent", "inactive"]
        direction:
          type: integer
        last_occurrence_bars_ago:
          type: integer
          nullable: true
        occurrences_last_100_bars:
          type: integer
        description:
          type: string
    PatternsAllResponse:
      type: object
      required: [symbol, tf, total_patterns, active, recent, inactive, patterns]
      properties:
        symbol:
          type: string
        tf:
          $ref: "#/components/schemas/Timeframe"
        total_patterns:
          type: integer
        active:
          type: integer
        recent:
          type: integer
        inactive:
          type: integer
        patterns:
          type: array
          items:
            $ref: "#/components/schemas/PatternAllItem"
...
    JobResponse:
      type: object
      required: [id, kind, status, created_at]
      properties:
        id:
          type: string
        kind:
          type: string
        payload:
          type: object
        status:
          type: string
          enum: ["pending", "running", "done", "failed"]
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        result:
          type: object
          nullable: true
        progress_current:
          type: integer
        progress_total:
          type: integer
        progress_message:
          type: string